requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.1
h11==0.14.0
idna==3.6
motor==3.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
